% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NRC.R
\name{nrc}
\alias{nrc}
\title{Cálculo de peso final mediante ecuaciones del NRC.}
\usage{
nrc(pv, cms, em, edad, gp, x = 0.0031, y = 1, data)
}
\arguments{
\item{pv}{(vector) Peso al iniciar la engorda.}

\item{cms}{(vector) Materia seca consumida.}

\item{x}{(vector) Dato establecido por NRC.}

\item{y}{(vector) Dato establecido por NRC.}

\item{data}{(vector) Datos reales para entrenamiento.}

\item{emta}{(vector) Energía metabolizable de la ración.}

\item{gdp}{(vector) Ganancia de peso diaria en kg.}
}
\value{
Peso al finalizar la engorda por NRC.
}
\description{
Cálculo de peso final mediante ecuaciones del NRC.
}
