% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLBOV.R
\name{rlb}
\alias{rlb}
\title{Regresión lineal para predecir pesos finales en bovinos de engorda.}
\usage{
rlb(pv, cms, em, gpd, pf, datos)
}
\arguments{
\item{em}{(vector) Energía metabolizable de la ración.}

\item{pf}{(vector) Pesos finales al termino de la engorda.}

\item{datos}{(vector) Datos reales para entrenamiento.}

\item{pi}{(vector) Peso al iniciar la engorda.}

\item{ms}{(vector) Materia seca consumida.}

\item{gdp}{(vector) Ganancia de peso diaria en kg.}
}
\value{
Peso al finalizar la engorda.
}
\description{
Regresión lineal para predecir pesos finales en bovinos de engorda.
}
